+++ src/Control/Monad/Apiary/Action/Internal.hs
@@ -10,6 +10,7 @@
 {-# LANGUAGE RecordWildCards #-}
 {-# LANGUAGE DeriveDataTypeable #-}
 {-# LANGUAGE TupleSections #-}
+{-# LANGUAGE DeriveFunctor #-}
 
 module Control.Monad.Apiary.Action.Internal where
 
@@ -53,7 +54,7 @@ data ApiaryConfig = ApiaryConfig
     , failStatus          :: Status
     , failHeaders         :: ResponseHeaders
       -- | used by 'Control.Monad.Apiary.Filter.root' filter.
-    , rootPattern         :: [S.ByteString]
+    , rootPattern         :: [T.Text]
     , mimeType            :: FilePath -> S.ByteString
     }
 
@@ -77,7 +78,7 @@ instance Default ApiaryConfig where
         , defaultHeaders      = []
         , failStatus          = internalServerError500
         , failHeaders         = []
-        , rootPattern         = ["", "/", "/index.html", "/index.htm"]
+        , rootPattern         = ["", "/", "index.html", "index.htm"]
         , mimeType            = defaultMimeLookup . T.pack
         }
 
@@ -88,16 +89,16 @@ data ActionState = ActionState
     , actionStatus   :: Status
     , actionHeaders  :: ResponseHeaders
     , actionReqBody  :: Maybe ([Param], [File])
-    , actionPathInfo :: [T.Text]
+    , actionFetches  :: [T.Text]
     }
 
-initialState :: ApiaryConfig -> Request -> ActionState
-initialState conf req = ActionState
+initialState :: ApiaryConfig -> ActionState
+initialState conf = ActionState
     { actionResponse = responseLBS (defaultStatus conf) (defaultHeaders conf) ""
     , actionStatus   = defaultStatus  conf
     , actionHeaders  = defaultHeaders conf
     , actionReqBody  = Nothing
-    , actionPathInfo = pathInfo req
+    , actionFetches  = []
     }
 {-# INLINE initialState #-}
 
@@ -110,37 +111,20 @@ data ActionEnv = ActionEnv
     }
 
 data Action a 
-    = Continue a
+    = Continue ActionState a
     | Pass
     | Stop Response
+    deriving (Functor)
 
-newtype ActionT m a = ActionT { unActionT :: forall b. 
+newtype ActionT m a = ActionT { runActionT ::
     ActionEnv
     -> ActionState
-    -> (a -> ActionState -> m (Action b))
-    -> m (Action b)
+    -> m (Action a)
     }
 
-runActionT :: Monad m => ActionT m a
-           -> ActionEnv -> ActionState
-           -> m (Action (a, ActionState))
-runActionT m env st = unActionT m env st $ \a st' ->
-    st' `seq` return (Continue (a, st'))
-{-# INLINE runActionT #-}
-
-actionT :: Monad m 
-        => (ActionEnv -> ActionState -> m (Action (a, ActionState)))
-        -> ActionT m a
-actionT f = ActionT $ \env st cont -> f env st >>= \case
-    Pass             -> return Pass
-    Stop s           -> return $ Stop s
-    Continue (a,st') -> st' `seq` cont a st'
-{-# INLINE actionT #-}
-
 -- | n must be Monad, so cant be MFunctor.
-hoistActionT :: (Monad m, Monad n)
-             => (forall b. m b -> n b) -> ActionT m a -> ActionT n a
-hoistActionT run m = actionT $ \e s -> run (runActionT m e s)
+hoistActionT :: (forall b. m b -> n b) -> ActionT m a -> ActionT n a
+hoistActionT run m = ActionT $ \e s -> run (runActionT m e s)
 {-# INLINE hoistActionT #-}
 
 execActionT :: ApiaryConfig -> Documents -> ActionT IO () -> Application
@@ -149,62 +133,71 @@ execActionT config doc m request send =
 #else
 execActionT config doc m request = let send = return in
 #endif
-    runActionT m (ActionEnv config request doc) (initialState config request) >>= \case
+    runActionT m (ActionEnv config request doc) (initialState config) >>= \case
 #ifdef WAI3
         Pass           -> notFound config request send
 #else
         Pass           -> notFound config request
 #endif
         Stop s         -> send s
-        Continue (_,r) -> send $ actionResponse r
+        Continue r _   -> send $ actionResponse r
 
 --------------------------------------------------------------------------------
 
-instance Functor (ActionT m) where
-    fmap f m = ActionT $ \env st cont ->
-        unActionT m env st (\a s' -> s' `seq` cont (f a) s')
+instance Functor m => Functor (ActionT m) where
+    fmap f m = ActionT $ \env st ->
+        fmap f <$> runActionT m env st
 
-instance Applicative (ActionT m) where
-    pure x = ActionT $ \_ st cont -> cont x st
-    mf <*> ma = ActionT $ \env st cont ->
-        unActionT mf env st  $ \f st'  ->
-        unActionT ma env st' $ \a st'' ->
-        st' `seq` st'' `seq` cont (f a) st''
-
+instance (Functor m, Monad m) => Applicative (ActionT m) where
+    pure x = ActionT $ \_ s -> return $ Continue s x
+    mf <*> ma = ActionT $ \env st ->
+        runActionT mf env st >>= \case
+            Pass   -> return Pass
+            Stop r -> return $ Stop r
+            Continue st' f -> runActionT ma env st' >>= \case
+                Continue st'' a -> return $ Continue st'' (f a)
+                Pass   -> return Pass
+                Stop r -> return $ Stop r
+            
 instance Monad m => Monad (ActionT m) where
-    return x = ActionT $ \_ st cont -> cont x st
-    m >>= k  = ActionT $ \env st cont ->
-        unActionT m env st $ \a st' ->
-        st' `seq` unActionT (k a) env st' cont
-    fail s = ActionT $ \(ActionEnv{actionConfig = c}) _ _ -> return $
-        Stop (responseLBS (failStatus c) (failHeaders c) $ LC.pack s)
+    return x = ActionT $ \_ st -> return $ Continue st x
+    m >>= k  = ActionT $ \env st ->
+        runActionT m env st >>= \case
+            Pass   -> return Pass
+            Stop r -> return $ Stop r
+            Continue st' a -> runActionT (k a) env st' >>= \case
+                Pass   -> return Pass
+                Stop r -> return $ Stop r
+                Continue st'' b -> return $ Continue st'' b
+    fail s = ActionT $ \ActionEnv{actionConfig = c} _ ->
+        return $ Stop (responseLBS (failStatus c) (failHeaders c) $ LC.pack s)
 
 instance MonadIO m => MonadIO (ActionT m) where
-    liftIO m = ActionT $ \_ st cont ->
-        liftIO m >>= \a -> cont a st
+    liftIO m = ActionT $ \_ st ->
+        Continue st `liftM` liftIO m
 
 instance MonadTrans ActionT where
-    lift m = ActionT $ \_ st cont ->
-        m >>= \a -> cont a st
+    lift m = ActionT $ \_ st ->
+        Continue st `liftM` m
 
 instance MonadThrow m => MonadThrow (ActionT m) where
-    throwM e = ActionT $ \_ st cont ->
-        throwM e >>= \a -> cont a st
+    throwM e = ActionT $ \_ st ->
+        Continue st `liftM` throwM e
 
 instance MonadCatch m => MonadCatch (ActionT m) where
-    catch m h = actionT $ \env st -> 
+    catch m h = ActionT $ \env st -> 
         catch (runActionT m env st) (\e -> runActionT (h e) env st)
     {-# INLINE catch #-}
 
 instance MonadMask m => MonadMask (ActionT m) where
-    mask a = actionT $ \env st ->
+    mask a = ActionT $ \env st ->
         mask $ \u -> runActionT (a $ q u) env st
       where
-        q u m = actionT $ \env st -> u (runActionT m env st)
-    uninterruptibleMask a = actionT $ \env st ->
+        q u m = ActionT $ \env st -> u (runActionT m env st)
+    uninterruptibleMask a = ActionT $ \env st ->
         uninterruptibleMask $ \u -> runActionT (a $ q u) env st
       where
-        q u m = actionT $ \env st -> u (runActionT m env st)
+        q u m = ActionT $ \env st -> u (runActionT m env st)
     {-# INLINE mask #-}
     {-# INLINE uninterruptibleMask #-}
 
@@ -215,11 +208,11 @@ instance (Monad m, Functor m) => Alternative (ActionT m) where
     {-# INLINE (<|>) #-}
 
 instance Monad m => MonadPlus (ActionT m) where
-    mzero = actionT $ \_ _ -> return Pass
-    mplus m n = actionT $ \e s -> runActionT m e s >>= \case
-        Continue a -> return $ Continue a
-        Stop stp   -> return $ Stop stp
-        Pass       -> runActionT n e s
+    mzero = ActionT $ \_ _ -> return Pass
+    mplus m n = ActionT $ \e s -> runActionT m e s >>= \case
+        Continue st a -> return $ Continue st a
+        Stop stp      -> return $ Stop stp
+        Pass          -> runActionT n e s
     {-# INLINE mzero #-}
     {-# INLINE mplus #-}
 
@@ -227,10 +220,10 @@ instance MonadBase b m => MonadBase b (ActionT m) where
     liftBase = liftBaseDefault
 
 instance MonadTransControl ActionT where
-    newtype StT ActionT a = StActionT { unStActionT :: Action (a, ActionState) }
-    liftWith f = actionT $ \e s -> 
-        liftM (\a -> Continue (a,s)) (f $ \t -> liftM StActionT $ runActionT t e s)
-    restoreT m = actionT $ \_ _ -> liftM unStActionT m
+    newtype StT ActionT a = StActionT { unStActionT :: Action a }
+    liftWith f = ActionT $ \e s -> 
+        liftM (\a -> Continue s a) (f $ \t -> liftM StActionT $ runActionT t e s)
+    restoreT m = ActionT $ \_ _ -> liftM unStActionT m
 
 instance MonadBaseControl b m => MonadBaseControl b (ActionT m) where
     newtype StM (ActionT m) a = StMT { unStMT :: ComposeSt ActionT m a }
@@ -244,7 +237,7 @@ instance MonadReader r m => MonadReader r (ActionT m) where
 --------------------------------------------------------------------------------
 
 getEnv :: Monad m => ActionT m ActionEnv
-getEnv = ActionT $ \e s c -> c e s
+getEnv = ActionT $ \e st -> return $ Continue st e
 
 -- | get raw request. since 0.1.0.0.
 getRequest :: Monad m => ActionT m Request
@@ -257,30 +250,30 @@ getDocuments :: Monad m => ActionT m Documents
 getDocuments = liftM actionDocuments getEnv
 
 getRequestBody :: MonadIO m => ActionT m ([Param], [File])
-getRequestBody = ActionT $ \e s c -> case actionReqBody s of
-    Just b  -> c b s
+getRequestBody = ActionT $ \e s -> case actionReqBody s of
+    Just b  -> return $ Continue s b
     Nothing -> do
         (p,f) <- liftIO $ P.parseRequestBody P.lbsBackEnd (actionRequest e)
         let b = (p, map convFile f)
-        c b s { actionReqBody = Just b }
+        return $ Continue s { actionReqBody = Just b } b
   where
     convFile (p, P.FileInfo{..}) = File p fileName fileContentType fileContent
 
 -- | parse request body and return params. since 0.9.0.0.
 getReqParams :: MonadIO m => ActionT m [Param]
-getReqParams = fst <$> getRequestBody
+getReqParams = fst `liftM` getRequestBody
 
 -- | parse request body and return files. since 0.9.0.0.
 getReqFiles :: MonadIO m => ActionT m [File]
-getReqFiles = snd <$> getRequestBody
+getReqFiles = snd `liftM` getRequestBody
 
 --------------------------------------------------------------------------------
 
 modifyState :: Monad m => (ActionState -> ActionState) -> ActionT m ()
-modifyState f = ActionT $ \_ s c -> c () (f s)
+modifyState f = ActionT $ \_ s -> return $ Continue (f s) ()
 
-getState :: ActionT m ActionState
-getState = ActionT $ \_ s c -> c s s
+getState :: Monad m => ActionT m ActionState
+getState = ActionT $ \_ s -> return $ Continue s s
 
 -- | set status code. since 0.1.0.0.
 status :: Monad m => Status -> ActionT m ()
@@ -314,11 +307,11 @@ contentType c = modifyHeader
 
 -- | stop handler and send current state. since 0.3.3.0.
 stop :: Monad m => ActionT m a
-stop = ActionT $ \_ s _ -> return $ Stop (actionResponse s)
+stop = ActionT $ \_ s -> return $ Stop (actionResponse s)
 
 -- | stop with response. since 0.4.2.0.
 stopWith :: Monad m => Response -> ActionT m a
-stopWith a = ActionT $ \_ _ _ -> return $ Stop a
+stopWith a = ActionT $ \_ _ -> return $ Stop a
 
 -- | redirect handler
 --
@@ -354,7 +347,7 @@ redirectPermanently = redirectWith movedPermanently301
 -- since 0.6.2.0.
 redirect :: Monad m => S.ByteString -> ActionT m ()
 redirect to = do
-    v <- httpVersion <$> getRequest
+    v <- httpVersion `liftM` getRequest
     if v == http11
         then redirectWith seeOther303 to
         else redirectWith status302   to
@@ -367,7 +360,7 @@ redirect to = do
 -- since 0.3.3.0.
 redirectTemporary :: Monad m => S.ByteString -> ActionT m ()
 redirectTemporary to = do
-    v <- httpVersion <$> getRequest
+    v <- httpVersion `liftM` getRequest
     if v == http11
         then redirectWith temporaryRedirect307 to
         else redirectWith status302            to
@@ -391,7 +384,7 @@ file' f p = rawResponse (\s h -> responseFile s h f p)
 -- | set response body file content and detect Content-Type by extension. since 0.1.0.0.
 file :: Monad m => FilePath -> Maybe FilePart -> ActionT m ()
 file f p = do
-    mime <- mimeType <$> getConfig
+    mime <- mimeType `liftM` getConfig
     contentType (mime f)
     file' f p